Move context related code from AblyProvider to new AblyContext file
