+++ src/platform/react-hooks/src/index.ts
@@ -6,3 +6,4 @@ export * from './AblyProvider.js';
 export * from './hooks/useChannelStateListener.js';
 export * from './hooks/useConnectionStateListener.js';
 export { ChannelProvider } from './ChannelProvider.js';
+export * from './AblyContext.js';
